india_health_atlas/requirements.txt
# India Health Atlas — Python dependencies
# Install all at once:  pip install -r requirements.txt

pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
plotly>=5.11.0
folium>=0.14.0
streamlit>=1.18.0
geopandas>=0.12.0
pyyaml>=6.0
openpyxl>=3.0.10
scikit-learn>=1.1.0